Added resource-recording with flag `record-resource-usage`
This will record a number of things. Have a look at the added object `ResourceUsageDto ` for the info written to a file `exec-logger/resource-usage.json`. This file gets appended to everytime.

There is a `DurationIncreaser` object to allow us to dynamically increase the sleep duration between checking for resource usage. Currently it will at the end wait a maximum of 30 seconds. This "increaser" has a `iterationsPerDuration` and `durationList`.

Currently the resource usage is only implemented for windows, see the `resource_usage/helper.go` file for the `VisitLinux` and `VisitDarwin` which has a TODO and panic.